Move shuttle to a different folder
