+++ client-example/client.sh
