+++ client-example/licenses.txt
