+++ common/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod comm;
+pub mod entry;
+pub mod game;
