+++ shuttle/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "fun-shuttle"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+fun-together = { path = "../common" }
+shuttle-runtime = { version = "0.13.0" }
+tokio = { version = "1", features = ["full"] }
